Optimise day 9 part 2
Cache the free spans in a dictionary of indices to run lengths, then update this each time a file is moved. Since every file is moved exactly once, from right to left, we never need to update the cache with the free space created where the file was. As a result, we can cache once and update as we go. This reduces run time of part 2 to approximately 6.5s from 16m.